Move astToHtml util method to AstHtmlVisitor class
